requirements.txt
asgiref==3.8.1
Django==4.2.3
django-ckeditor==6.7.1
drf_yasg==1.21.7
django-js-asset==2.2.0
django-modeltranslation==0.19.3
djangorestframework==3.15.2
pillow==10.3.0
sqlparse==0.5.0
